requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary==2.9.5
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
google-auth>=2.25.0
python-dotenv>=1.0.0
skyfield>=1.46
google-generativeai>=0.3.0
pytz>=2023.3
